Streamly.FSNotify

Description

Introduction

This provides file watching as a Streamly stream. You can either watch recursively (namely, a directory's contents and all its subdirectories as well), or not. You can also filter out file system events you are not interested in. Lastly, we provide a compositional scheme for constructing filters for file system events.

Example

This example program watches /home/koz/c-project (and any of its subdirectories) for added or modified files with a .c extension, and emits the change to the terminal, along with a timestamp of when it happened, forever:

import System.FilePath ((</>))
import Streamly.FSNotify (EventPredicate, hasExtension, isDirectory, invert, isDeletion, conj, watchTree)
import qualified Streamly.Prelude as SP

-- conj -> both must be true
-- invert -> true when the argument would be false and vice versa
isCSourceFile :: EventPredicate
isCSourceFile = hasExtension "c" `conj` invert isDirectory

notDeletion :: EventPredicate
notDeletion = invert isDeletion

srcPath :: FilePath
srcPath = "home" </> "koz" </> "c-project"

-- first value given by watchTree stops the watcher
-- we don't use it here, but if you want to, just call it
main :: IO ()
main = do
    (_, stream) <- watchTree srcPath $ isCSourceFile `conj` notDeletion
    SP.drain . SP.mapM go $ stream
  where
    go = \case
        Added p t _ -> putStrLn $ "Created: " ++ show p ++ " at " ++ show t
        Modified p t _ -> putStrLn $ "Modified: " ++ show p ++ " at " ++ show t
        _ -> pure ()

Basic types

data FSEntryType

Allows us to designate Events as being fired by a directory or a non-directory entry.

Constructors

Dir

NotDir

data Event

A file system notification.

Constructors

Added FilePath UTCTime FSEntryType
Creation event

Modified FilePath UTCTime FSEntryType
Modification event

Removed FilePath UTCTime FSEntryType
Deletion event

Other FilePath UTCTime Text
Some other kind of event

data StopWatching m

An object, which, when executed with stopWatching, stops a file system watch.

eventPath :: Event -> FilePath

Helper to retrieve the file path associated with an event.

eventTime :: Event -> UTCTime

Helper to retrieve an event's timestamp.

eventFSEntry :: Event -> Maybe FSEntryType

Helper to retrieve whether the event stems from a directory or not. Returns Nothing if the event is not 'basic' (that is, not a creation, modification or deletion).

Events and predicates

newtype EventPredicate

A 'test' for whether we want to 'notice' an event. Should return True for events we care about.

Constructors

EventPredicate

Fields

runPredicate :: Event -> Bool

everything :: EventPredicate

The trivial predicate (allows any event through).

nothing :: EventPredicate

The null predicate (allows no events through).

isDirectory :: EventPredicate

Allows through events that are caused by directories. Note that this will assume that non-'basic' events (that is, not creations, modifications or deletions) do not stem from directories; use with care.

hasExtension :: FilePath -> EventPredicate

Allows through events triggered by file system entries with a specific extension.

isCreation :: EventPredicate

Allows through only creation events.

isModification :: EventPredicate

Allows through only modification events.

isDeletion :: EventPredicate

Allows through only deletion events.

isBasic :: EventPredicate

Allows through only 'basic' events (namely creation, modification and deletion).

invert :: EventPredicate -> EventPredicate

'Flips' the predicate - what it used to allow through is now blocked, and vice versa.

conj :: EventPredicate -> EventPredicate -> EventPredicate

Predicate conjunction (meaning that both have to be true for the result to be true). Synonym for times.

disj :: EventPredicate -> EventPredicate -> EventPredicate

Predicate disjunction (meaning that either has to be true for the result to be true). Synonym for plus.

Watchers

watchDirectory :: (IsStream t, MonadAsync m) => FilePath -> EventPredicate -> m (StopWatching m, t m Event)

Watch a given directory, but only at one level (thus, subdirectories will not be watched recursively).

watchDirectoryWith :: (IsStream t, MonadAsync m) => WatchConfig -> FilePath -> EventPredicate -> m (StopWatching m, t m Event)

As watchDirectory, but with a specified set of watch options.

watchTree :: (IsStream t, MonadAsync m) => FilePath -> EventPredicate -> m (StopWatching m, t m Event)

Watch a given directory recursively (thus, subdirectories will also have their contents watched).

watchTreeWith :: (IsStream t, MonadAsync m) => WatchConfig -> FilePath -> EventPredicate -> m (StopWatching m, t m Event)

As watchTree, but with a specified set of watch options.